TT#14008 don't delete calls from foreign DB during restore
If a keyspace notification SET is received and the call already exists
as a foreign call, the call is first destroyed before being re-restored.
The call destruction involves a DEL from Redis on the "hosted DB"
number, which points to the foreign DB. This makes it impossible to then
restore the call because it's just been deleted.

closes #1308

Change-Id: Ie895b021441b2d299f8ebb5bde1824b01e12633c